Fix imports and tests after rebase
